+++ internal/bundler/bundler_default_test.go
@@ -2788,7 +2788,41 @@ func TestUseStrictDirectiveBundleIssue1837(t *testing.T) {
 	})
 }
 
-func TestUseStrictDirectiveBundleIssue2264(t *testing.T) {
+func TestUseStrictDirectiveBundleIIFEIssue2264(t *testing.T) {
+	default_suite.expectBundled(t, bundled{
+		files: map[string]string{
+			"/entry.js": `
+				'use strict'
+				export let a = 1
+			`,
+		},
+		entryPaths: []string{"/entry.js"},
+		options: config.Options{
+			Mode:          config.ModeBundle,
+			AbsOutputFile: "/out.js",
+			OutputFormat:  config.FormatIIFE,
+		},
+	})
+}
+
+func TestUseStrictDirectiveBundleCJSIssue2264(t *testing.T) {
+	default_suite.expectBundled(t, bundled{
+		files: map[string]string{
+			"/entry.js": `
+				'use strict'
+				export let a = 1
+			`,
+		},
+		entryPaths: []string{"/entry.js"},
+		options: config.Options{
+			Mode:          config.ModeBundle,
+			AbsOutputFile: "/out.js",
+			OutputFormat:  config.FormatCommonJS,
+		},
+	})
+}
+
+func TestUseStrictDirectiveBundleESMIssue2264(t *testing.T) {
 	default_suite.expectBundled(t, bundled{
 		files: map[string]string{
 			"/entry.js": `