+++ CHANGELOG.md
@@ -13,6 +13,22 @@
     * `.woff` => `font/woff`
     * `.woff2` => `font/woff2`
 
+* Remove `"use strict";` when targeting ESM ([#2347](https://github.com/evanw/esbuild/issues/2347))
+
+    All ES module code is automatically in strict mode, so a `"use strict";` directive is unnecessary. With this release, esbuild will now remove the `"use strict";` directive if the output format is ESM. This change makes the generated output file a few bytes smaller:
+
+    ```js
+    // Original code
+    'use strict'
+    export let foo = 123
+
+    // Old output (with --format=esm --minify)
+    "use strict";let t=123;export{t as foo};
+
+    // New output (with --format=esm --minify)
+    let t=123;export{t as foo};
+    ```
+
 ## 0.14.47
 
 * Make global names more compact when `||=` is available ([#2331](https://github.com/evanw/esbuild/issues/2331))